[r=newell-jensen][bug=1666997][author=ltrager] Notify and warn the user when the ephemeral operating system is unavailable.
The settings page allows a user to select which operating system MAAS should use during commissioning and other ephemeral environments. This list was built from operating systems that are currently available on rack controllers. If the currently selected operating system is not available the first choice was selected. This led the user to believe a usable operating system was selected when it really wasn't.

When the ephemeral operating system is not available MAAS creates a notification for all users, shows an error on the images page, lists the currently selected commissioning operating system as unavailable on the settings page, and an exception is raised when the user tries to use the ephemeral operating. Users are no longer able to remove the current ephemeral operating system from the BootSourceSelection table.