[r=newell-jensen][bug=1688009][author=ltrager] Ignore timeout when telling the rack to disable and stop itself on deletion.
When the system running a rack controller is shutdown and a user tries to
delete it the region may still think the rack is connected. The region will
attempt to send an RPC call instructing the rack to disable and stop itself.
This is to ensure the rack doesn't reconnect causing MAAS to recreate a node
object for it. If the system is turned off it won't be coming back. Wait for
the timeout to occur but shorten it as we really do want the rackd service to
be off after deletion.