[r=andreserl][bug=][author=blake-rouse] Add new 'maas init', 'maas config', 'maas status' commands for the MAAS snap.
This branch adds the new 'maas init', 'maas config', 'maas status' commands for the MAAS snap. The MAAS snap now starts in 'none' mode with nothing configured. The 'maas init' gives an interactive setup mode to configure the snap, 'maas config' allow the user to get the current configuration or change the snap configuration, and the 'maas status' provides the current status of all services running in the snap.