[r=mpontillo][bug=1276945][author=bjornt] Add support for using the http proxy as a peer proxy.
If use_peer_proxy is True in the settings, the built-in http proxy will
be configured to use the specified http proxy as a cache peer. So that
machines will go through the built-in proxy, which in turn will go
through the external http proxy to download packages.

It's possible to configure use_peer_proxy via the API for now. The UI
will come later.