Refactor rack registration to allow deferred fabric creation.
This allows update_interfaces() to be called in two modes: one which
creates all fabrics immediately (the current behavior), and the other
which only creates 'disconnected' physical interfaces and skips child
interface creation and deletion.

This work is necessary so that in the future, a rack can register,
perform beaconing, and update its interfaces with any additional
topology data the beacons can provide.

Also updates all test scenarios for update_interfaces() to make sure
they work with the old 'one phase' approach and the new 'two phase'
approach. (That is, the interfaces are created in exactly the same
way even if fabric creation is deferred.)