Install packages before running scripts.
Scripts may now set the packages attribute. This attribute is a dictionary
which describes what apt packages, snap packages, or URLs to download. Package
installation happens before the script runs and is not counted against its
runtime. The region will be notified when script installation starts and an
installation failure will set the script to failed installation.
maas-run-remote-scripts will export the environment variable DOWNLOAD_PATH
which is the path to all URLs downloaded. Any URL containing a tar or zip will
 be extracted and any deb or snap file will be installed.