Add a switches tab to the nodes listing.
This is a very basic listing, which only shows the name and the owner of the node. Later on, the listing will be improved, and actions like delete/deploy will be added, as well as having the listing get updated when nodes change in the backend.

Since the switches work is still early and experimental, you need to manually append '?switches=on' to the URL to see the new UI.