Add an index on the created field on the events table.
Since the order of events on a node cannot be determined by its ID, the created field is used to sort the events in the correct order. Without an index on this field it makes the query take a long time. This adds the index to help the sorting of events by created time.