Add volume_groups, raids, cache_sets, and bcaches field to the Machine API output.
This allows the user to see the different storage items that exist for the machine. The user can then use the provided ID to load that object over the API. It does not provide the full object as that would make the API return to much data for a machines list or machines read.