Fixes LP: #1726474 - Retry FOREIGN_KEY_VIOLATIONS in the transactional retry context.
FOREIGN_KEY_VIOLATIONS can occur do to other regiond processes performing an action on the same node. This will now retry those actions if they fail with FOREIGN_KEY_VIOLATIONS. If the code is wrong in general this will not mask that error as after the retries pass it will still fail.