Enabled builtin Django connection pooling that improves MAAS performance.
Previously a new connection would be created every time an action need to be performed with the database. This now keeps a pool of Postgresql and reuses those connections instead of creating a new connect every time (which is very slow). A new database_conn_max_age configuration option has been added that can be adjusted in different environments if needed.

The default is 5 minutes. Every 5 minutes the connection will be recycled and a new one created.