Have MAAS control the workers and divide the work between a master process and the worker processes.
This is a big change in the way MAAS workers, as the master process now controls the workers. An IPC communication is created between the master at the workers. At the moment that IPC is very simple, but will grow as the postgres listener and the RPC connection work is moved from each worker to run in the master and a proxy in the workers.