Capture the NIC vendor, product, and firmware version.
MAAS captures the network information about a system by processing the output
of ip addr. ip addr does not contain the vendor, product, or firmware version
of the NIC but it is contained in the lshw output. Since the lshw
commissioning script is sent before ip addr output it can be processed by the
ip addr hook to gain the additional information.