Fix test__calls_handler_with_update_on_update intermittent failure.
Explicitly pass a changed attribute. Before we relied on that the django
object wasn't exactly the same as the database row, which wasn't always
the case. So even though django force a database update, if no values
had changed, postgres didn't consider it an update, so the trigger
wasn't called.