Factor out common init code.
We currenty have 'maas init' in the snap, and we want to add add a 'maas
init' command to the deb package as well.  This branch factors out the
functionality we want in the deb package, so that the snap and deb
command can share the code.

Also, tests were added for the factored out code, since most of it was
untested before.