LP: #1760037 - Only load requested data from the GeneralManager.
Previously all data was loaded by the GeneralManager on page load. This
caused pages that only needed a few items from the GeneralManager to
request data it did not need. To retrieve a reference to the data from
the GeneralManager GeneralManager.getData() must be used. This is done
before manager load and now flags the item as requested. On
GeneralManager load only items which have been flaged requested are
now loaded.