Ensure the services are running once its connected to the region. Don't cache the service status rack side.
Having the connection before checking services ensures that the current service status will be reported to the region. Remove the caching of services on the rack because the region can become out of sync from the rack and they rack will never tell the region about the running services because its cached in the rack.