LP: #1761600 - Don't return deferLater value.
In f8bf8cb MAAS was modified to use asyncio or uvloop as the backend for
twisted. The value of deferLater was returned in maybe_change_power_state
which fixes a number of unit tests. Doing this causes the defer to be
resolved with the caller instead of running later. This meant when a user
issues a power action, such as deploy or release, the websocket and
subsequent RPC call didn't resolve until the power action is completed.

This removes the return so the power action can happen later greatly
speeding up power actions. The unit tests are fixed by patching
deferLater in the tests to use maybeDeferred.