LP: #1768323 - Create generic link for old style hwe kernels.
18b0508 made sure only the ga kernel is launched when the ephemerial
environment is started is a generic kernel by making sure only the 'generic'
hard link is only created for the ga kernel. This did not account for old
style kernels using the format hwe-<first letter of release>. This caused no
generic hard link to be created. Trusty deployments were not effected as MAAS
specifies the real kernel and subarch name(hwe-t). This does not happen with
precise as the commissioning OS is used to deploy and generic is used for the
image subarch instead.

Also fixes a bug where MAAS would use theimage URL for any subarch matching
the arch, osystem, and series. This allows Precise to deploy with current
streams.