Upgrade from using default_storage_pool power parameter to using the new default_storage_pool foreign key set on the Pod.
This allows the old path to work by handling the default_storage_pool set on the power_parameters to still propagate to the rack controller. The new way overrides this setting on compose and on refresh of the pod the setting in power_parameters in the pod is migrated to the model.