Rename Pod.default_pool to Pod.pool.
It was originaly default_pool, since pools are tied to permissions, but pods
weren't covered. Now that pools don't deal with permissions, it was decided to
name it pool, since pods will most be part of the RBAC work later.

I renamed the DB patches, since we want to backport this change.