Retrieve list of power types directly from the PowerDriverRegistry
Previously MAAS would obtain the list of all power drivers available on rack controllers. This list was obtained from each rack controller which queried the PowerDriverRegistry. The PowerDriverRegistry is available directly to the region. PowerDrivers cannot be added or removed without modifications to the MAAS code base. Remove the RPC call to the rack controllers and query the PowerDriverRegistry directly from the region.