RPC connections between the rack and region controllers cause a new rpc.state file to be updated in /var/lib/maas. This file holds the URLs that the rack controller could use to talk to the region controllers. When the rack controller cannot reach any region controller defined in rackd.conf, the rpc.state file will be used to try and make a connection to the region controllers.
This is allows multiple region controllers to exist for a rack controller, but only requiring 1 to be specified in the rackd.conf. The rack controller will automatically learn about the other region controllers and store that information. Next time when the rack controller needs to contact the region controller to get the RPC endpoint information it will fallback to the other known region controllers when the one specified in the rackd.conf is not responding.