Add timeout on the rack controller of 30 seconds for DHCP configuration.
This ensures that even if the DHCP configuration gets stuck that at most 30 seconds later it will be cancelled and another one will be performed. The timeout error is propagated back to the region controller so the region will request the rack controller to update again, with the latest information.