Calculate a better interface name based on the specified labels when composing a machine.
This is required because Juju passes in numeric interface labels, such as '0'. MAAS uses these labels (which are strings) to create correspondingly-named interfaces on the composed machine.

The algorithm used to calculate the ifname from the label was adapted from the Juju bridge name algorithm (which is used when a bridge name ends up being more than 15 characters).

Partially fixes bug #1789495.