Make pxeconfig.cfg/ an optional part of the bootload config path.
The builtin PXE bootloader on s390x KVM sets the default bootloader path
prefix to pxelinux.cfg/ while on other platforms pxelinux.cfg/ is always part
of the path. Make pxelinux.cfg/ an optional part of the bootloader config
path so booting works now and will continue to work if this is ever
corrected.