Sync the resource pools to the RBAC service when RBAC is enabled and changes occur that require a sync.
Only when the RBAC service is enabled the RegionControllerService will sync the resource pools to the RBAC micro-service. Any time a change occurs that requires a sync, that will also sync the information to the RBAC micro-service. Locking is used to ensure that only one region controller will sync the information to the RBAC service.