LP: #1795051 - Add MegaRAID support to smartctl tests.
smartctl can gather SMART data and run tests for individual drives on a
hardware RAID controller if given the device type and bus id. To determine
the bus id the proprietary tool storcli from LSI must be used. As storcli
is proprietary and unavailable in the Ubuntu archives or as a Snap MAAS is
unable to automatically install it. If another test installs storcli onto
the machine smartctl tests will use it to test all drives on all MegaRAID controllers. If storcli is not found the test will continue to be marked
as skipped.