LP: #1799862 - Handle lsblk errors and hangs in maas-run-remote-scripts
maas-run-remote-scripts uses lsblk to map storage devices that MAAS wants
tested to what they are on the booted system. If lsblk failed to run the
thread executing the test would never finish resulting in
maas-run-remote-scripts hanging. If lsblk fails to run or takes longer
than 60 seconds maas-run-remote-scripts will mark all tests which use a
storage parameter as a failure. The test log will show the failure occurred
because of lsblk.