LP: #1800233 - Deprecate Trusty as a commissioning series.
Storage tests do not work when using Trusty due to maas-run-remote-scripts
using an unsupported lsblk option. The stress-ng and NTP tests have never
worked when using Trusty. As Trusty is going EOL in April 2019 it makes
sense to deprecate Trusty as a commissioning series now. Trusty will still
be used when deploying Trusty.