Add Node model method to copy storage from a source node on the node. Requirements for the operation to succeed is that the nodes have the same number of disks and that the destination nodes disks are the same size or bigger than the source node.
Matching of the block devices between the 2 nodes goes through a hierarchy to provide the best match between the 2 nodes. A three step process is done to ensure the best match, while also increasing the likely hood of the clone operation succeeding.

Due to the nature of virtual block devices that allow more partitions and filesystems to be layered on top, the code is rather complex. It iterates over the filesystem groups and cache sets in layers cloning each layer until moving up to the next layer to clone. This cloning of filesystem groups and cache sets is continued until no more filesystem groups or cache sets exist for the node.