LP: #1821605 - Drop amtterm as a requirement for AMT power control
amtterm (which provided amttool) was previously used for really old AMT implementations, and MAAS used to rely on it to discover the version of AMT to decide whether to use wsmancli or amttool. Nowadays, however, wsmancli is used for the discovery process as well as it being the default power management version.

That said, MAAS is failing to power manage a machine if *only* wsmancli is installed (but not amtterm). Since amtterm is really deprecated from the AMT perspective (but we keep it in MAAS just in case), dropping the requirement for it being installed would allow to power manage AMT machines only by having wsmancli installed.