LP: #1825244 - Send WS what storage layout has been detected on the machine.
The websocket node handler will now include a new 'detected_storage_layout'
field. This field is the storage layout name that matches the currently
applied storage configuration closest. This does not imply that the layout
was previously applied, a user may have manually configured storage to be
similar to an existing layout. This check matches on the base configuration,
users may have modified the configuration. For example if the LVM layout was
applied and a user reconfigures the system to have one EXT4 partition on the
first disk and a second on a second disk this will return 'flat'.